Enhance MCP data processing with API database integration and refactor build tools
Two big changes in this commit:
1. Processing of data happens in memory. JSON is retired. Data is deserialized in single binary
2. the API DB and instructions are now embedded in the MCP executable.

This commit introduces significant updates to the MCP data processing framework:

- Added a new `api_database` module containing data types and models for the MCP server, improving organization and maintainability.
- Refactored the `build_api_definitions` and `build_embeddings` modules to utilize the new `api_database` types, streamlining data handling.
- Updated `Cargo.toml` files across various modules to include dependencies on `api_database`, ensuring consistent access to shared types.
- Removed deprecated `main.rs` files from `build_api_definitions` and `build_embeddings`, transitioning to library-based structures for better modularity.
- Enhanced the `data_transformer` to support unified API database generation, including optional JSON export for debugging.

These changes collectively improve the efficiency, usability, and maintainability of the MCP data processing pipeline, providing a more robust framework for users interacting with the API.
