+++ mcp/api_database/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "api_database"
+version = "0.1.0"
+edition = "2021"
+description = "API data types and model for vim_rs MCP server"
+
+[lib]
+name = "api_database"
+path = "src/lib.rs"
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+chrono = { version = "0.4", features = ["serde"] }
+indexmap = { version = "2.0", features = ["serde"] }
+anyhow = "1.0"
+tracing = "0.1"
+
+# For JSON loading (optional, used by legacy load_from_dir)
+serde_json = { version = "1.0", optional = true }
+
+[features]
+default = []
+json = ["serde_json"]
+
