+++ mcp/data_processing/build_api_definitions/Cargo.toml
@@ -3,13 +3,13 @@ name = "build-api-definitions"
 version = "0.1.0"
 edition = "2021"
 
-[[bin]]
-name = "build-api-definitions"
-path = "src/main.rs"
+[lib]
+name = "build_api_definitions"
+path = "src/lib.rs"
 
 [dependencies]
 vim_build = { path = "../../../vim_build" }
-vim_mcp_server = { path = "../../server" }
+api_database = { path = "../../api_database" }
 chrono = { version = "0.4", features = ["serde"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
@@ -18,5 +18,4 @@ thiserror = "2.0"
 convert_case = "0.8"
 check_keyword = "0.3"
 tracing = "0.1"
-tracing-subscriber = "0.3"
