+++ mcp/data_processing/build_api_definitions/src/lib.rs
@@ -1,8 +1,11 @@
-pub mod json_emitter;
+pub mod api_builder;
 
 use anyhow::Result;
 use std::path::Path;
 use vim_build::{load_openapi, load_vim_model};
+
+// Re-export types for external use
+pub use api_builder::ApiDefinitions;
 
 /// Types whose hierarchies are pruned from the model. The listed types will be generated. Their
 /// descendants will not. This means that:
@@ -17,14 +20,27 @@ use vim_build::{load_openapi, load_vim_model};
 /// correct type_name_ for the descendant types.
 pub const PRUNED_TYPES: [&str; 2] = ["MethodFault", "Event"];
 
-/// Build API definitions from OpenAPI specification
+/// Build API definitions from OpenAPI specification and return in-memory data.
+/// This is the primary function for the data pipeline.
+pub fn build_api_data(vi_json_spec_path: &Path) -> Result<ApiDefinitions> {
+    // Load OpenAPI spec and transform to vim_model
+    let openapi = load_openapi(vi_json_spec_path)?;
+    let model = load_vim_model(&openapi, Some(&PRUNED_TYPES))?;
+
+    // Build all API definitions in memory
+    api_builder::build_api_data(&model)
+        .map_err(|e| anyhow::anyhow!("Failed to build API data: {}", e))
+}
+
+/// Build API definitions and write to JSON files (for debugging).
+/// This is the legacy function that writes JSON files.
 pub fn build_api_definitions(vi_json_spec_path: &Path, output_dir: &Path) -> Result<()> {
     // Load OpenAPI spec and transform to vim_model
     let openapi = load_openapi(vi_json_spec_path)?;
     let model = load_vim_model(&openapi, Some(&PRUNED_TYPES))?;
 
     // Generate JSON files
-    json_emitter::emit_mcp_data(&model, output_dir, &PRUNED_TYPES)
+    api_builder::emit_mcp_data(&model, output_dir, &PRUNED_TYPES)
         .map_err(|e| anyhow::anyhow!("Failed to emit MCP data: {}", e))?;
 
     Ok(())
