+++ mcp/data_processing/build_embeddings/Cargo.toml
@@ -3,20 +3,15 @@ name = "build-embeddings"
 version = "0.1.0"
 edition = "2021"
 
-[[bin]]
-name = "build-embeddings"
-path = "src/main.rs"
+[lib]
+name = "build_embeddings"
+path = "src/lib.rs"
 
 [dependencies]
-vim_mcp_server = { path = "../../server" }
+api_database = { path = "../../api_database" }
 fastembed = "5.2.0"
-tokio = { version = "1.42", features = ["full"] }
 anyhow = "1.0"
 tracing = "0.1"
-tracing-subscriber = "0.3"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-bincode = "1.3"
 
 # CUDA GPU acceleration (optional)
 ort = { version = "2.0.0-rc.7", optional = true, features = ["cuda", "copy-dylibs"] }
