+++ mcp/data_processing/build_examples/Cargo.toml
@@ -3,14 +3,14 @@ name = "build-examples"
 version = "0.1.0"
 edition = "2021"
 
-[[bin]]
-name = "build-examples"
-path = "src/main.rs"
+[lib]
+name = "build_examples"
+path = "src/lib.rs"
 
 [dependencies]
+api_database = { path = "../../api_database" }
 anyhow = "1.0"
 tracing = "0.1"
-tracing-subscriber = "0.3"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 walkdir = "2.5"
