+++ mcp/server/Cargo.toml
@@ -4,15 +4,16 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+# API database types (shared with build tools)
+api_database = { path = "../api_database" }
+
 tokio = { version = "1.42", features = ["full"] }
 serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
 anyhow = "1.0"
 tracing = "0.1"
 tracing-subscriber = "0.3"
 phf = "0.11"
 thiserror = "1.0"
-chrono = { version = "0.4", features = ["serde"] }
 indexmap = { version = "2.0", features = ["serde"] }
 
 # MCP protocol implementation
@@ -20,15 +21,16 @@ rmcp = { version = "0.8.5", features = ["server", "macros"] }
 schemars = { version = "1.0", features = ["derive"] }
 tokio-util = "0.7.17"
 
-# Embeddings and vector search (optional for semantic search)
-fastembed = { version = "5.2.0", optional = true }
-rayon = { version = "1.10", optional = true }
-bincode = { version = "1.3", optional = true }
+# Binary serialization (required for embedded database)
+bincode = "1.3"
 
-# CUDA GPU acceleration (requires CUDA 12.x + cuDNN 9.x runtime)
+# Embeddings and vector search (core functionality)
+fastembed = "5.2.0"
+rayon = "1.10"
+
+# CUDA GPU acceleration (optional, requires CUDA 12.x + cuDNN 9.x runtime)
 ort = { version = "2.0.0-rc.7", optional = true, features = ["cuda", "copy-dylibs"] }
 
 [features]
-default = ["embeddings"]
-embeddings = ["fastembed", "rayon", "bincode"]
-cuda = ["embeddings", "ort"]
+default = []
+cuda = ["ort"]
